Todo, Done 이미지 추가 및 설명 삽입
