+++ package.json
@@ -12,7 +12,8 @@
     "clsx": "^2.1.1",
     "next": "15.1.3",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-responsive": "^10.0.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
